+++ src/Http/Controllers/DocumentsController.php
@@ -1,34 +1,35 @@
 <?php
 
+
 namespace MBober35\Fileable\Http\Controllers;
 
-use App\Http\Controllers\Controller;
+
 use App\Models\File;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
-use MBober35\Fileable\Events\ChangeImageOrder;
-use MBober35\Fileable\Events\ImageChanged;
-use MBober35\Fileable\Facades\GalleryActions;
+use MBober35\Fileable\Events\ChangeDocumentOrder;
+use MBober35\Fileable\Events\DocumentChanged;
+use MBober35\Fileable\Facades\DocumentActions;
 
-class GalleryController extends Controller
+class DocumentsController
 {
     protected $modelObj;
 
     public function __construct()
     {
-        $route = \request()->route();
+        $route = request()->route();
         if (! empty($route)) {
             $model = $route->parameter("model", false);
             $id = $route->parameter("id", false);
             if ($model && $id) {
-                $this->modelObj = GalleryActions::getGalleryModel($model, $id);
+                $this->modelObj = DocumentActions::getModel($model, $id);
             }
         }
     }
 
     /**
-     * Получить изображения.
+     * Получить документы.
      *
      * @param string $model
      * @param int $id
@@ -39,80 +40,58 @@ class GalleryController extends Controller
         return response()
             ->json([
                 "success" => true,
-                "images" => GalleryActions::getGalleryResource($this->modelObj),
+                "documents" => DocumentActions::getResource($this->modelObj),
             ]);
     }
 
-    /**
-     * Загрузить изображения.
-     *
-     * @param Request $request
-     * @param string $model
-     * @param int $id
-     * @return \Illuminate\Http\JsonResponse
-     * @throws \Illuminate\Validation\ValidationException
-     */
     public function store(Request $request, string $model, int $id)
     {
         $this->storeValidator($request->all());
-
-        $mime = $request->file("image")->getClientOriginalExtension();
+        // Получить расширение.
+        $mime = $request->file("document")->getClientOriginalExtension();
+        // Сохранить файл.
         $fileName = Str::random(40) . "." . $mime;
-        $path = $request->file("image")->storeAs("gallery/$model", $fileName);
+        $path = $request->file("document")->storeAs("documents/{$model}", $fileName);
         $name = $request->get("name");
-        $type = "image";
-        $image = File::create(
+        $type = "document";
+        $doc = File::create(
             compact("path", "name", "mime", "type")
         );
-        $this->modelObj->images()->save($image);
-        $image = GalleryActions::setPriority($image);
-        ImageChanged::dispatch($image, "created", $this->modelObj);
+        $this->modelObj->documents()->save($doc);
+        $doc = DocumentActions::setPriority($doc);
+        DocumentChanged::dispatch($doc, "created", $this->modelObj);
 
         return response()
             ->json([
                 "success" => true,
-                "images" => GalleryActions::getGalleryResource($this->modelObj),
+                "documents" => DocumentActions::getResource($this->modelObj),
             ]);
     }
 
-    /**
-     * @param array $data
-     * @throws \Illuminate\Validation\ValidationException
-     */
     protected function storeValidator(array $data)
     {
         Validator::make($data, [
-            "image" => ["required", "image"],
+            "document" => ["required", "file"],
             "name" => ["required", "string", "max:100"]
         ], [
-            "image.required" => "Файл не найден",
-            "image.image" => "Файл должен быть изображением",
+            "document.required" => "Файл не найден !!",
+            "document.file" => "Не удалось загрузить документ",
         ], [
             "name" => "Имя",
             "image" => "Файл",
         ])->validate();
     }
 
-    /**
-     * Обновить название.
-     *
-     * @param Request $request
-     * @param string $model
-     * @param int $id
-     * @param File $file
-     * @return \Illuminate\Http\JsonResponse
-     * @throws \Illuminate\Validation\ValidationException
-     */
     public function update(Request $request, string $model, int $id, File $file)
     {
         $this->updateValidator($request->all());
         $file->name = $request->get("name");
         $file->save();
-        ImageChanged::dispatch($file, "updated", $this->modelObj);
+        DocumentChanged::dispatch($file, "updated", $this->modelObj);
         return response()
             ->json([
                 "success" => true,
-                "images" => GalleryActions::getGalleryResource($this->modelObj),
+                "documents" => DocumentActions::getResource($this->modelObj),
             ]);
     }
 
@@ -129,36 +108,24 @@ class GalleryController extends Controller
         ])->validate();
     }
 
-    /**
-     * Изменить порядок вывода.
-     *
-     * @param Request $request
-     * @param string $model
-     * @param int $id
-     * @return \Illuminate\Http\JsonResponse
-     * @throws \Illuminate\Validation\ValidationException
-     */
     public function order(Request $request, string $model, int $id)
     {
         $this->orderValidator($request->all());
-
-        $ids = $request->get("images");
+        $ids = $request->get("documents");
         foreach ($ids as $priority => $id) {
             try {
                 $file = File::find($id);
                 $file->priority = $priority;
                 $file->save();
-            }
-            catch (\Exception $exception) {
+            } catch (\Exception $exception) {
                 continue;
             }
         }
-        ChangeImageOrder::dispatch($this->modelObj);
-
+        ChangeDocumentOrder::dispatch($this->modelObj);
         return response()
             ->json([
                 "success" => true,
-                "images" => GalleryActions::getGalleryResource($this->modelObj),
+                "documents" => DocumentActions::getResource($this->modelObj),
             ]);
     }
 
@@ -169,29 +136,20 @@ class GalleryController extends Controller
     protected function orderValidator(array $data)
     {
         Validator::make($data, [
-            "images" => ["required", "array"]
+            "documents" => ["required", "array"]
         ], [], [
-            "images" => "Изображения",
+            "documents" => "Документы",
         ])->validate();
     }
 
-    /**
-     * Удаление изображения.
-     *
-     * @param string $model
-     * @param int $id
-     * @param File $file
-     * @return \Illuminate\Http\JsonResponse
-     * @throws \Exception
-     */
     public function destroy(string $model, int $id, File $file)
     {
         $file->delete();
-        ImageChanged::dispatch($file, "deleted", $this->modelObj);
+        DocumentChanged::dispatch($file, "deleted", $this->modelObj);
         return response()
             ->json([
                 "success" => true,
-                "images" => GalleryActions::getGalleryResource($this->modelObj)
+                "documents" => DocumentActions::getResource($this->modelObj),
             ]);
     }
 }