+++ src/ServiceProvider.php
@@ -7,9 +7,11 @@ use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\ServiceProvider as BaseProvider;
 use MBober35\Fileable\Commands\FilesCommand;
 use MBober35\Fileable\Commands\ThumbnailsClearCommand;
+use MBober35\Fileable\Helpers\DocumentActionsManager;
 use MBober35\Fileable\Helpers\GalleryActionsManager;
 use MBober35\Fileable\Helpers\ThumbnailActionsManager;
 use MBober35\Fileable\Observers\FileOberver;
+use MBober35\Fileable\View\Components\Documents;
 use MBober35\Fileable\View\Components\Gallery;
 
 class ServiceProvider extends BaseProvider
@@ -36,10 +38,16 @@ class ServiceProvider extends BaseProvider
         $this->app->singleton("thumbnail-actions", function () {
             return new ThumbnailActionsManager;
         });
+        $this->app->singleton("document-actions", function () {
+            return new DocumentActionsManager;
+        });
 
         // Конфигурация.
         $this->mergeConfigFrom(
             __DIR__ . "/config/gallery.php", "gallery"
+        );
+        $this->mergeConfigFrom(
+            __DIR__ . "/config/documents.php", "documents"
         );
     }
 
@@ -63,11 +71,13 @@ class ServiceProvider extends BaseProvider
 
         // Адреса.
         $this->loadRoutesFrom(__DIR__ . '/routes/gallery.php');
+        $this->loadRoutesFrom(__DIR__ . '/routes/documents.php');
         $this->loadRoutesFrom(__DIR__ . '/routes/thumb.php');
 
         // Конфигурация.
         $this->publishes([
             __DIR__ . "/config/gallery.php" => config_path("gallery.php"),
+            __DIR__ . "/config/documents.php" => config_path("documents.php"),
         ], "config");
 
         // Подключение шаблонов.
@@ -75,6 +85,7 @@ class ServiceProvider extends BaseProvider
 
         // Компоненты.
         Blade::component("gallery", Gallery::class);
+        Blade::component("documents", Documents::class);
     }
 
     /**