+++ src/config/documents.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    "models" => [],
+    "resource" => \MBober35\Fileable\Http\Resources\DocumentResource::class,
+];