+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+## [1.2.0]
+- Добавлен трейт `ShouldDocuments`
+- Добавлен компонент `x-documents`
+
+### Обновление
+- php artisan fileable --no-replace
+- php artisan vendor:publish --provider="MBober35\Fileable\ServiceProvider" --tag=public
+
 ## [1.1.0]
 - Добавлен трейт `ShouldDocument`
 - Изменено сохранение изображений